+++ scripts/nordic-publish.ts
@@ -13,7 +13,7 @@ import FtpClient from 'ftp';
 import semver from 'semver';
 import calculateShasum from 'shasum';
 
-import { AppJson, PackageJson, SourceJson } from '../src/utils/AppTypes';
+import { AppInfo, PackageJson, SourceJson } from '../src/utils/AppTypes';
 import checkAppProperties from './check-app-properties';
 
 interface LegacyAppInfo {
@@ -36,7 +36,7 @@ interface App {
     version: string;
     shasum: string;
     sourceUrl: string;
-    appJsonName: string;
+    appInfoName: string;
     releaseNotesFilename: string;
     iconFilename: string;
     isOfficial: boolean;
@@ -149,7 +149,7 @@ const packOrReadPackage = (options: Options): App => {
         shasum,
         sourceUrl: options.sourceUrl,
         isOfficial: options.deployOfficial,
-        appJsonName: `${name}.json`,
+        appInfoName: `${name}.json`,
         releaseNotesFilename: `${name}-Changelog.md`,
         iconFilename: `${name}.svg`,
         packageJson,
@@ -303,7 +303,7 @@ const getUpdatedSourceJson = async (app: App): Promise<SourceJson> => {
         name: sourceJson.name,
         apps: [
             ...new Set(sourceJson.apps).add(
-                `${app.sourceUrl}/${app.appJsonName}`
+                `${app.sourceUrl}/${app.appInfoName}`
             ),
         ].sort(),
     };
@@ -311,10 +311,8 @@ const getUpdatedSourceJson = async (app: App): Promise<SourceJson> => {
 
 const downloadExistingVersions = async (app: App) => {
     try {
-        const appJson = <AppJson>(
-            JSON.parse(await downloadFileContent(app.appJsonName))
-        );
-        return appJson.versions;
+        const appInfoContent = await downloadFileContent(app.appInfoName);
+        return (JSON.parse(appInfoContent) as AppInfo).versions;
     } catch (error) {
         console.log(
             `No previous app versions found due to: ${errorAsString(error)}`
@@ -330,7 +328,7 @@ const failBecauseOfMissingProperty = () => {
     );
 };
 
-const getUpdatedAppJson = async (app: App): Promise<AppJson> => {
+const getUpdatedAppInfo = async (app: App): Promise<AppInfo> => {
     const versions = await downloadExistingVersions(app);
 
     const { name, displayName, description, homepage, version } =
@@ -363,10 +361,10 @@ const uploadSourceJson = (sourceJson: SourceJson) =>
         'source.json'
     );
 
-const uploadAppJson = (app: App, appJson: AppJson) =>
+const uploadAppInfo = (app: App, appInfo: AppInfo) =>
     uploadFile(
-        Buffer.from(JSON.stringify(appJson, undefined, 2)),
-        app.appJsonName
+        Buffer.from(JSON.stringify(appInfo, undefined, 2)),
+        app.appInfoName
     );
 
 const uploadPackage = (app: App) => uploadFile(app.filename, app.filename);
@@ -416,13 +414,13 @@ const main = async () => {
 
         const legacyAppInfo = await getUpdatedLegacyAppInfo(app);
         const sourceJson = await getUpdatedSourceJson(app);
-        const appJson = await getUpdatedAppJson(app);
+        const appInfo = await getUpdatedAppInfo(app);
 
         await uploadChangelog(app);
         await uploadIcon(app);
         await uploadPackage(app);
         await uploadLegacyAppInfo(app, legacyAppInfo);
-        await uploadAppJson(app, appJson);
+        await uploadAppInfo(app, appInfo);
         await uploadSourceJson(sourceJson);
 
         console.log('Done');